+++ src/main/java/com/example/Blog/App/Api/payload/PostDto.java
@@ -7,9 +7,9 @@ import lombok.Data;
 
 @Data
 public class PostDto {
-    private Long id;
+    private Integer id;
     private String title;
     private String description;
     private String content;
-    private Long categoryId;
+    private Integer categoryId;
 }
